Change content-resources groupId from fi.nls.oskari to org.oskari
