+++ content-resources/pom.xml
@@ -8,6 +8,7 @@
 		<artifactId>oskari-parent</artifactId>
         <version>2.0.0-SNAPSHOT</version>
 	</parent>
+    <groupId>org.oskari</groupId>
     <artifactId>content-resources</artifactId>
     <packaging>jar</packaging>
     <name>Content Resources</name>
