+++ content-resources/src/main/java/flyway/analysis/V1_0_4__add_baselayer.java
@@ -1,12 +1,10 @@
 package flyway.analysis;
 
-import fi.nls.oskari.geoserver.GeoserverPopulator;
-import fi.nls.oskari.geoserver.LayerHelper;
+import org.oskari.usercontent.GeoserverPopulator;
+import org.oskari.usercontent.LayerHelper;
 import fi.nls.oskari.util.PropertyUtil;
 import org.flywaydb.core.api.migration.BaseJavaMigration;
 import org.flywaydb.core.api.migration.Context;
-
-import java.sql.Connection;
 
 public class V1_0_4__add_baselayer extends BaseJavaMigration {
