+++ content-resources/src/main/java/flyway/analysis/V1_0_8__migrate_style_to_json.java
@@ -2,11 +2,9 @@ package flyway.analysis;
 
 import fi.nls.oskari.log.LogFactory;
 import fi.nls.oskari.log.Logger;
-import fi.nls.oskari.util.UserDataStyleMigrator;
+import org.oskari.usercontent.UserDataStyleMigrator;
 import org.flywaydb.core.api.migration.BaseJavaMigration;
 import org.flywaydb.core.api.migration.Context;
-
-import java.sql.Connection;
 
 public class V1_0_8__migrate_style_to_json extends BaseJavaMigration {
     private static final Logger LOG = LogFactory.getLogger(V1_0_8__migrate_style_to_json.class);
