+++ control-base/pom.xml
@@ -25,6 +25,10 @@
         </dependency>
 
         <dependency>
+            <groupId>org.oskari</groupId>
+            <artifactId>service-admin</artifactId>
+        </dependency>
+        <dependency>
             <groupId>fi.nls.oskari.service</groupId>
             <artifactId>oskari-myplaces</artifactId>
         </dependency>
