+++ control-base/src/main/java/fi/nls/oskari/control/layer/SaveLayerHandler.java
@@ -5,7 +5,6 @@ import fi.mml.map.mapwindow.service.wms.WebMapService;
 import fi.mml.map.mapwindow.service.wms.WebMapServiceParseException;
 import fi.mml.map.mapwindow.util.OskariLayerWorker;
 import fi.nls.oskari.annotation.OskariActionRoute;
-import fi.nls.oskari.cache.JedisManager;
 import fi.nls.oskari.control.*;
 import fi.nls.oskari.domain.map.DataProvider;
 import fi.nls.oskari.domain.map.OskariLayer;
@@ -23,14 +22,12 @@ import fi.nls.oskari.service.capabilities.OskariLayerCapabilitiesHelper;
 import fi.nls.oskari.util.*;
 import org.oskari.admin.LayerCapabilitiesHelper;
 import org.oskari.log.AuditLog;
-import fi.nls.oskari.wfs.GetGtWFSCapabilities;
 import fi.nls.oskari.wmts.WMTSCapabilitiesParser;
 import fi.nls.oskari.wmts.domain.WMTSCapabilities;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.oskari.permissions.model.*;
 import org.oskari.service.util.ServiceFactory;
-import org.oskari.service.wfs3.WFS3Service;
 
 import javax.servlet.http.HttpServletRequest;
 import java.net.MalformedURLException;
