--- geoserver-ext/geoserver-rest-client/src/main/java/fi/nls/oskari/geoserver/Workspace.java
@@ -1,11 +1,0 @@
-package fi.nls.oskari.geoserver;
-
-import com.fasterxml.jackson.annotation.JsonRootName;
-
-/**
- * Created by SMAKINEN on 1.9.2015.
- */
-@JsonRootName("workspace")
-public class Workspace {
-    public String name;
-}
