+++ geoserver-ext/geoserver-rest-client/src/main/java/org/oskari/usercontent/Namespace.java
@@ -1,0 +1,12 @@
+package org.oskari.usercontent;
+
+import com.fasterxml.jackson.annotation.JsonRootName;
+
+/**
+ * Created by SMAKINEN on 1.9.2015.
+ */
+@JsonRootName("namespace")
+public class Namespace {
+    public String prefix;
+    public String uri;
+}
