+++ geoserver-ext/geoserver-rest-client/src/main/java/org/oskari/usercontent/Workspace.java
@@ -1,0 +1,11 @@
+package org.oskari.usercontent;
+
+import com.fasterxml.jackson.annotation.JsonRootName;
+
+/**
+ * Created by SMAKINEN on 1.9.2015.
+ */
+@JsonRootName("workspace")
+public class Workspace {
+    public String name;
+}
