+++ geoserver-ext/geoserver-rest-client/src/test/resources/org/oskari/usercontent/default.sld
