+++ service-csw/pom.xml
@@ -17,8 +17,12 @@
 
     <dependencies>
         <dependency>
-            <groupId>fi.nls.oskari</groupId>
-            <artifactId>content-resources</artifactId>
+            <groupId>fi.nls.oskari.service</groupId>
+            <artifactId>oskari-map</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>fi.nls.oskari.service</groupId>
+            <artifactId>oskari-search</artifactId>
         </dependency>
         <dependency>
             <groupId>org.mybatis</groupId>
