+++ service-webapp/pom.xml
@@ -16,7 +16,7 @@
 
     <dependencies>
         <dependency>
-            <groupId>fi.nls.oskari</groupId>
+            <groupId>org.oskari</groupId>
             <artifactId>content-resources</artifactId>
         </dependency>
         <dependency>
